+++ features/nanostack/FEATURE_NANOSTACK/nanostack-interface/NanostackInterface.cpp
@@ -32,12 +32,6 @@
 #include "ns_trace.h"
 #define TRACE_GROUP "nsif"
 
-#include "nanostack-event-loop/eventOS_scheduler.h"
-
-#define nanostack_lock()            eventOS_scheduler_mutex_wait()
-#define nanostack_unlock()          eventOS_scheduler_mutex_release()
-#define nanostack_assert_locked()   //MBED_ASSERT(eventOS_scheduler_mutex_is_owner())
-
 #define NS_INTERFACE_SOCKETS_MAX  16  //same as NanoStack SOCKET_MAX
 #define NANOSTACK_SOCKET_UDP 17 // same as nanostack SOCKET_UDP
 #define NANOSTACK_SOCKET_TCP 6  // same as nanostack SOCKET_TCP
@@ -443,82 +437,6 @@ void NanostackSocket::event_connnect_closed(socket_callback_t *sock_cb)
     MBED_ASSERT((SOCKET_MODE_STREAM == mode) ||
                 (SOCKET_MODE_CONNECTING == mode));
     close();
-}
-
-MeshInterfaceNanostack::MeshInterfaceNanostack()
-    : phy(NULL), _network_interface_id(-1), _device_id(-1), eui64(),
-      ip_addr_str(), mac_addr_str(), connect_semaphore(0)
-{
-    // Nothing to do
-}
-
-MeshInterfaceNanostack::MeshInterfaceNanostack(NanostackRfPhy *phy)
-    : phy(phy), _network_interface_id(-1), _device_id(-1), connect_semaphore(0)
-{
-    // Nothing to do
-}
-
-nsapi_error_t MeshInterfaceNanostack::initialize(NanostackRfPhy *phy)
-{
-    mesh_system_init();
-    if (this->phy != NULL) {
-        error("Phy already set");
-    }
-    this->phy = phy;
-    return 0;
-}
-
-void MeshInterfaceNanostack::mesh_network_handler(mesh_connection_status_t status)
-{
-    nanostack_lock();
-
-    if (status == MESH_CONNECTED) {
-        connect_semaphore.release();
-    }
-
-    nanostack_unlock();
-}
-
-nsapi_error_t MeshInterfaceNanostack::register_rf()
-{
-    nanostack_lock();
-
-    _device_id = phy->rf_register();
-    if (_device_id < 0) {
-        nanostack_unlock();
-        return -1;
-    }
-    // Read mac address after registering the device.
-    phy->get_mac_address(eui64);
-    sprintf(mac_addr_str, "%02x:%02x:%02x:%02x:%02x:%02x:%02x:%02x", eui64[0], eui64[1], eui64[2], eui64[3], eui64[4], eui64[5], eui64[6], eui64[7]);
-
-    nanostack_unlock();
-
-    return 0;
-}
-
-NetworkStack * MeshInterfaceNanostack::get_stack()
-{
-    return NanostackInterface::get_stack();
-}
-
-const char *MeshInterfaceNanostack::get_ip_address()
-{
-    nanostack_lock();
-
-    const char *ret = NULL;
-    if (getOwnIpAddress(ip_addr_str, sizeof ip_addr_str)) {
-        ret = ip_addr_str;
-    }
-
-    nanostack_unlock();
-
-    return ret;
-}
-
-const char *MeshInterfaceNanostack::get_mac_address()
-{
-    return mac_addr_str;
 }
 
 NanostackInterface * NanostackInterface::_ns_interface;
